[REFACTOR] [Issue17] observer 구조 통일
