+++ app/src/main/java/com/example/android_repo_04/utils/Event.kt
@@ -1,6 +1,7 @@
-package com.example.android_repo_04.view
+package com.example.android_repo_04.utils
 
 import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.Observer
 
 open class Event<out T>(private val content: T) {
 
@@ -18,6 +19,15 @@ open class Event<out T>(private val content: T) {
 
     fun peekContent(): T = content
 }
+
+class EventObserver<T>(private val onEventUnhandledContent: (T) -> Unit) : Observer<Event<T>> {
+    override fun onChanged(event: Event<T>?) {
+        event?.getContentIfNotHandled()?.let { value ->
+            onEventUnhandledContent(value)
+        }
+    }
+}
+
 fun MutableLiveData<Event<Unit>>.emit() {
     value = Event(Unit)
 }
