+++ app/src/main/java/com/example/android_repo_04/view/main/notification/NotificationFragment.kt
@@ -13,7 +13,8 @@ import com.example.android_repo_04.R
 import com.example.android_repo_04.api.GitHubApiRepository
 import com.example.android_repo_04.data.dto.notification.Notification
 import com.example.android_repo_04.databinding.FragmentNotificationBinding
-import com.example.android_repo_04.view.Event
+import com.example.android_repo_04.utils.Event
+import com.example.android_repo_04.utils.EventObserver
 import com.example.android_repo_04.viewmodel.CustomViewModelFactory
 import com.example.android_repo_04.viewmodel.MainViewModel
 
@@ -43,7 +44,7 @@ class NotificationFragment: Fragment(), NotificationSwipeListener {
         }
     }
 
-    private val notificationRefreshEventObserver: (Event<Unit>) -> Unit = {
+    private val notificationRefreshEventObserver: (Unit) -> Unit = { event ->
         viewModel.requestNotifications()
     }
 
@@ -83,7 +84,7 @@ class NotificationFragment: Fragment(), NotificationSwipeListener {
     private fun observeData() {
         viewModel.notifications.observe(viewLifecycleOwner, notificationObserver)
         viewModel.readNotification.observe(viewLifecycleOwner, readNotificationObserver)
-        viewModel.notificationRefreshEvent.observe(viewLifecycleOwner, notificationRefreshEventObserver)
+        viewModel.notificationRefreshEvent.observe(viewLifecycleOwner, EventObserver(notificationRefreshEventObserver))
     }
 
     private fun setItemTouchHelper() {
