+++ app/src/main/java/com/example/android_repo_04/view/search/SearchActivity.kt
@@ -1,18 +1,15 @@
 package com.example.android_repo_04.view.search
 
 import android.os.Bundle
-import android.view.KeyEvent
 import android.view.View
-import android.view.inputmethod.EditorInfo
-import android.widget.TextView
 import androidx.appcompat.app.AppCompatActivity
 import androidx.lifecycle.ViewModelProvider
 import com.example.android_repo_04.R
 import com.example.android_repo_04.api.GitHubApiRepository
-import com.example.android_repo_04.data.dto.profile.User
 import com.example.android_repo_04.data.dto.search.Search
 import com.example.android_repo_04.databinding.ActivitySearchBinding
-import com.example.android_repo_04.view.Event
+import com.example.android_repo_04.utils.Event
+import com.example.android_repo_04.utils.EventObserver
 import com.example.android_repo_04.viewmodel.CustomViewModelFactory
 import com.example.android_repo_04.viewmodel.SearchViewModel
 
@@ -43,31 +40,25 @@ class SearchActivity : AppCompatActivity(), SearchRefreshListener {
         }
     }
 
-    private val clickEventObserver: (Event<Int>) -> Unit = { event ->
-        when(event.getContentIfNotHandled()) {
+    private val clickEventObserver: (Int) -> Unit = { event ->
+        when(event) {
             R.id.img_search_back -> onBackPressed()
             R.id.img_search_cancel -> binding.editSearchSearch.setText("")
         }
     }
 
-    private val refreshEventObserver: (Event<Unit>) -> Unit = { event ->
-        if(event.getContentIfNotHandled() != null) {
-            binding.editSearchSearch.run {
-                if(this.text.isEmpty() || this.text.isBlank()) {
-                    binding.refreshSearch.isRefreshing = false
-                } else {
-                    getSearchItems(this.text.toString())
-                }
+    private val refreshEventObserver: (Unit) -> Unit = { event ->
+        binding.editSearchSearch.run {
+            if(this.text.isEmpty() || this.text.isBlank()) {
+                binding.refreshSearch.isRefreshing = false
+            } else {
+                getSearchItems(this.text.toString())
             }
         }
     }
 
     private val searchAdapter: SearchAdapter by lazy {
         SearchAdapter(viewModel)
-    }
-
-    private val imgSearchClickListener: (View) -> Unit = {
-        getSearchItems(binding.editSearchSearch.text.toString())
     }
 
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -103,8 +94,8 @@ class SearchActivity : AppCompatActivity(), SearchRefreshListener {
     private fun observeData() {
         viewModel.searchItems.observe(this, searchItemsObserver)
         viewModel.searchText.observe(this, searchTextObserver)
-        viewModel.clickEvent.observe(this, clickEventObserver)
-        viewModel.refreshEvent.observe(this, refreshEventObserver)
+        viewModel.clickEvent.observe(this, EventObserver(clickEventObserver))
+        viewModel.refreshEvent.observe(this, EventObserver(refreshEventObserver))
     }
 
     private fun getSearchItems(query: String) {
