+++ app/src/main/java/com/example/android_repo_04/viewmodel/LoginViewModel.kt
@@ -1,12 +1,14 @@
 package com.example.android_repo_04.viewmodel
 
-import android.util.Log
-import androidx.lifecycle.*
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.viewModelScope
 import com.example.android_repo_04.BuildConfig
 import com.example.android_repo_04.api.GitHubLoginRepository
 import com.example.android_repo_04.utils.DataResponse
-import com.example.android_repo_04.view.Event
-import com.example.android_repo_04.view.emit
+import com.example.android_repo_04.utils.Event
+import com.example.android_repo_04.utils.emit
 import kotlinx.coroutines.launch
 
 class LoginViewModel(private val gitHubLoginRepository: GitHubLoginRepository) : ViewModel() {
